Modified readData.cs to work with ReadDetail page
Attend to modify the readData.cs to work with the new ReadDetail page
prev/next navigation.  Currently, the readData readMsg = new readData();
line will creates a new object of the readData which create a new msg
list.  What I want it to do is use the same list created in the
ReadLongListPage and select the coresponding msg from the messgeGroupt
list.  I guess this will not be  a problem once we get the correct
readData.cs in place which getting the msg from DB.