Make the UI more user friendly and new app icons
Fully supported the prev and next button on the ReadDetailsPage.  The
user can now not only going back and forth on pages but also roll back
to the first one when the last one is on the current view and going to
the last page if the first one is on the current view.
Changed ReadDetailsPage to go back ReadLongListPage when user click the
HW back button.  And, when the user click the HW back button on
ReadLongListPage, the app will exit instead of going back to previous
viewed pages.
Customized all the App icons to more look and feel for this app.
Added support for the About page for this app.